improvements around UI, DI, logging
